🎉 feat(rs_kmeans): Initial commit for Rust K-Means implementation This commit adds the initial files for a Rust implementation of the K-Means clustering algorithm. The implementation includes a `k_means` module and a `solver` module. The `k_means` module contains functions for calculating the mean of a vector of numbers and the distance between two vectors. The `solver` module will contain the implementation of the K-Means algorithm.
The `Cargo.toml` file is also included, which specifies the package name, version, and dependencies.

A simple `Hello, world!` program is included in the `main.rs` file, along with some unit tests for the `mean` and `distance` functions.

The `/target` directory is added to the `.gitignore` file to exclude build artifacts.

Signed-off-by: Sabry Awad <dr.sabry1997@gmail.com>